Refactors local storage code
- moves local storage code from movie-modal into a separate file
